refactor(geo_foundation): move foundation traits to geometry namespace
- Move core_foundation.rs to src/geometry/foundation/core_foundation.rs
- Move extension_foundation.rs to src/geometry/foundation/extension_foundation.rs
- Create geometry/foundation/mod.rs with re-exports
- Update geometry/mod.rs to include foundation module
- Update lib.rs: remove old module declarations, add backward compatibility aliases
- All imports automatically work due to public re-exports
- All 314 tests pass; structure now logically organized
