+++ model/geo_foundation/src/geometry/foundation/core_foundation.rs
