+++ model/geo_foundation/src/geometry/foundation/mod.rs
@@ -1,0 +1,9 @@
+//! Geometry Foundation - 幾何形状の基盤トレイト定義
+//!
+//! Core Foundation と Extension Foundation の統合モジュール
+
+pub mod core_foundation;
+pub mod extension_foundation;
+
+pub use core_foundation::*;
+pub use extension_foundation::*;
