+++ model/geo_foundation/src/geometry/mod.rs
@@ -4,3 +4,4 @@
 
 pub mod core;
 pub mod extensions;
+pub mod foundation;
